do not update pitch on a sound if it has been stopped
If pitch update are allowed on stopped notes, it may happen that a track
changes pitch back to center value after a bend up/down which is then
audible as a 'snap back' while the note is fading out.

To avoid this we prevent altering the notes if it was stopped. For very
high depth pitch LFOs this may cause odd fadeouts if the release time is
set to very long. But hopefully this doesn't occur in practice.