fix profiles not saving properly
Silly bug with rather complicated cause.
Because during profile loading a std::move is performed, but the type Profile
has a custom copy constructor defined, the std::move won't do the
intended thing. More specifically that means in our case that the 'path'
member is not copied, which will cause all later saving attempts to
fail.