Rework songtable scanner (fixes #83)
The song table scanning is now performed per profile instead of only
once. This makes it possible to partially scan ROMs. Previously, when a
profile would manually override parts of the config (i.e. if the scanner
wouldn't find it), agbplay wouldn't load them. Even if the provided
values and offsets were all correct. This should be fixed now.