Improve songtable scanning (fixes #55)
In case of Gem Smashers, FindPlayerTable would sometimes find songthing
that looks like tables. However, if it is not possible to find the sound
mode or anything else, perhaps we haven't found the right table after
all.

This commit makes the scanner keep on scanning until it reaches the end
of the file.

Also, a bug is fixed where the scanner doesn't continue scanning the
exactly after the right song table candidate.